Separate dependencies that prevent building for Windows
