+++ cmd/desync/mount-index_windows.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"context"
+	"fmt"
+)
+
+func mountIdx(context.Context, []string) error {
+	return fmt.Errorf("Subcommand not available on this platform")
+}
