+++ mount-index.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package desync
 
 import (
